C++ code to swap LinkedList nodes in pair until we reach the end of the list
